LangChain_Prompts/requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-huggingface
huggingface-hub
transformers
